+++ examples/react/src/components/Connected.tsx
@@ -44,9 +44,6 @@ const checkoutProvider = searchParams.get('checkoutProvider')
 const onRampProvider = searchParams.get('onRampProvider')
 const checkoutPreset = searchParams.get('checkoutPreset') || 'forte-transak-payment-erc1155-sale-native-token-testnet'
 
-// @ts-ignore
-const isDev = __SEQUENCE_WEB_SDK_IS_DEV__
-
 export const Connected = () => {
   const { openTransactionStatusModal } = useTransactionStatusModal()
   const [isOpenCustomCheckout, setIsOpenCustomCheckout] = React.useState(false)