+++ examples/react/src/components/CustomCheckout/index.tsx
@@ -22,7 +22,6 @@ export const CustomCheckout = () => {
   const salesContractAddress = '0xe65b75eb7c58ffc0bf0e671d64d0e1c6cd0d3e5b'
   const collectionAddress = '0xdeb398f41ccd290ee5114df7e498cf04fac916cb'
   const price = '20000'
-  const contractId = '674eb5613d739107bbd18ed2'
 
   const chainId = 137