+++ packages/checkout/src/hooks/useTransakWidgetUrl.ts
@@ -1,6 +1,7 @@
-import { useQuery } from '@tanstack/react-query'
-import { getTransakWidgetUrl, type TransakWidgetUrlArgs } from '../api/data.js'
 import { useConfig } from '@0xsequence/hooks'
+import { useQuery } from '@tanstack/react-query'
+
+import { getTransakWidgetUrl, type TransakWidgetUrlArgs } from '../api/data.js'
 
 export const useTransakWidgetUrl = (args: TransakWidgetUrlArgs, disabled?: boolean) => {
   const { env, projectAccessKey } = useConfig()