+++ packages/checkout/src/utils/transak.ts
@@ -1,8 +1,6 @@
 import { ChainId } from '@0xsequence/network'
 import { zeroAddress } from 'viem'
 
-import type { AddFundsSettings } from '../contexts/AddFundsModal.js'
-
 export const TRANSAK_PROXY_ADDRESS = '0x4a598b7ec77b1562ad0df7dc64a162695ce4c78a'
 
 interface CountriesResult {