+++ packages/checkout/src/views/AddFunds.tsx
@@ -1,6 +1,5 @@
 import { Spinner, Text } from '@0xsequence/design-system'
-import { useAPIClient } from '@0xsequence/hooks'
-import React, { useEffect, useRef } from 'react'
+import { useEffect, useRef } from 'react'
 
 import { HEADER_HEIGHT } from '../constants/index.js'
 import type { AddFundsSettings } from '../contexts/AddFundsModal.js'