+++ packages/connect/src/styles.ts
@@ -2513,4 +2513,6 @@ export const styles = String.raw`
       --tw-gradient-to-position: 100%;
     }
   }
-}`
+}
+
+`